Date Night Calculator
Create a date night calculator. The user selects the costs of travel, dinner, dessert, and entertainment, the percentage of tip if needed, and inputs the tax rate. The program calculates total cost by adding the travel cost, the dinner cost, the dessert cost, the entertainment cost, the taxes on the dinner and dessert, and the tip on dinner cost with its associated taxes. The output is the total cost of the date.